Refactor how the program gets delivered to the frontend. Begin refactoring to support linking ProgramPersons to UserProfiles.
Now, to get program information on a page, you need to go through auth.programCache - and if you pass a reference to yourself, then live query should magically work. There should no longer be any reason to query the program directly - put everything through the cache
instead and you won't need to worry about processing updates. This should also cut down on whole-app re-renders (which can contribute to the UI flickering).

To receive live query updates, store a list of program info (e.g. of ProgramItem) in a state variable with a name matching the table name plus 's', e.g .called ProgramItems (or ProgramPersons or whatever). When there are updates, the cache will call setState on you. If you are storing a single item (e.g. a ProgramItem or ProgramPerson), store it in a state variable called the table name (e.g. ProgramItem), and setState will be magically called on you when the data changes.

See ProgramPersonDispaly, ProgramItemDisplay for examples of consuming a single program piece, see Program.js for an example of consuming the whole program.