#42 - Put in all of the scaffolding and basic UI to have users claim their author profiles. Display a user's program items in their tooltips, show the user info when hovering over an author that's linked to a user. Requires re-deploy of cloud functions, and add a `programPersons` array to UserProfile, `userProfile` pointer to ProgramPerson, and `programItems` array to Program Person.
The semantics are extremely loose from a security perspective: You can claim any author account, but when people see that you are the "author" of an item, it will show both the author name and the user's profile name. We can restrict the write access to this mapping to admins, but I don't think it's worth the dev time to implement that interface yet.

TBD - upload poster for programitems that you are marked as author of.

TBD - add you to the correct chat channel when you assume authorship.