Fixes #36. Set `overflow-y` to `visible` for bottom text-chat window.
The behaviour of `overflow` with `display: flex` is not particularly stable. In Firefox and (some versions of) Edge, `overflow-y: auto` does not result in a scroll bar. By forcing the overflow to be visible, it at least always shows the scrollbar (at the expense of showing it when it is not needed i.e. in chats with very few messages).