JS -> TS: ChatClient
Here be (small) dragons. This class relied heavily on JS idioms, so adding types was no small feat. There are some major issues with control flow that should probably be addressed in this class. For example:
- It imports and calls `setState` on a number of external React components.
- There are strange anti-patterns with promises (see `rhsChatPromise`.
- Some dead code that seems like it might still be important.

Since this class is at the core of the chat system, I plan on spending more time on this class tomorrow.