Simplistic asynchronous cache initialisation
I have POC that this caching strategy works. If you set up a conference Id in the session and reload the page, the conference name renders before the response from Parse comes back (by about 200ms on my machine) - so by the time you account for all the cache data download, this is a big saving.