[login]: Implement login via email properly (and more efficiently).
'Find by email' is not permitted by User table ACLs so we can't just lookup the email, then log in by username. It turns out, the Parse Server backend supports login via email, they just haven't included it the version of the JS SDK we're using (not sure when it'll land). So I've implemented it manually based on the proposed solution on their repo.